fix: グループ選択を /quiz/ インデックスルートに統合
group-selection.tsx の実装を quiz/index.tsx に移動し、
ボトムタブバーの /quiz リンクから直接アクセス可能にした
